prism_shot_folder_builder/requirements.txt
altgraph==0.17.4
natsort==8.4.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
tk==0.1.0